// apps/desktop/src/ipc/generated/AiHardwareDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AiGpuDto } from "./AiGpuDto";
import type { AiProfileDto } from "./AiProfileDto";
import type { AiRecommendationDto } from "./AiRecommendationDto";

export type AiHardwareDto = { memoryBytes: number, availableMemoryBytes: number, processor: string, processorThreads: number, graphics: Array<AiGpuDto>, freeDiskBytes: number | null, recommendation: AiRecommendationDto, profiles: Array<AiProfileDto>, };